Clean Code - Move ApiFeedScopedId from adapters to model.
