+++ src/main/java/org/opentripplanner/api/adapters/AgencyAndIdAdapter.java
@@ -1,5 +1,6 @@
 package org.opentripplanner.api.adapters;
 
+import org.opentripplanner.api.model.ApiFeedScopedId;
 import org.opentripplanner.model.FeedScopedId;
 
 import javax.xml.bind.annotation.adapters.XmlAdapter;
